fzy/requirements.txt
absl-py==2.1.0
accelerate==1.6.0
aiofiles==23.2.1
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.6.0
apted==1.0.3
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==2.4.1
async-lru==2.0.4
async-timeout==4.0.3
attrs==24.2.0
audioread==3.0.1
autocommand==2.2.2
av==13.1.0
babel==2.16.0
backports.tarfile==1.2.0
beautifulsoup4==4.12.3
bitsandbytes==0.44.1
black==24.10.0
bleach==6.2.0
boto3==1.35.5
botocore==1.35.5
Brotli 
certifi
cffi==1.17.1
cfgv==3.4.0
chardet==5.2.0
charset-normalizer
click==8.1.7
colorama==0.4.6
comm==0.2.2
contourpy==1.3.0
cycler==0.12.1
dataclasses-json==0.6.7
DataProperty==1.0.1
datasets==3.0.1
debugpy==1.8.7
decorator==5.1.1
decord==0.6.0
deepspeed==0.15.2
defusedxml==0.7.1
dill==0.3.8
Distance==0.1.3
distlib==0.3.9
distro==1.9.0
docstring_parser==0.16
editdistance==0.8.1
ego4d==1.7.3
einops==0.8.0
einops-exts==0.0.4
et-xmlfile==1.1.0
evaluate==0.4.3
exceptiongroup==1.2.2
executing==2.1.0
fastapi==0.115.2
fastjsonschema==2.20.0
ffmpy==0.4.0
filelock
fire==0.7.0
flash_attn==2.6.3
fonttools==4.54.1
fqdn==1.5.1
frozenlist==1.4.1
fsspec==2024.6.1
fvcore==0.1.5.post20221221
gmpy2 
GPUtil==1.4.0
gradio==4.44.1
gradio_client==1.3.0
grpcio==1.67.0
h11==0.14.0
hf-xet==1.1.4
hf_transfer==0.1.8
hjson==3.1.0
httpcore==1.0.6
httpx==0.27.2
huggingface-hub==0.31.2
identify==2.6.1
idna 
importlib_metadata==8.0.0
importlib_resources==6.4.0
inflect==7.3.1
iopath==0.1.10
ipdb==0.13.13
ipykernel==6.29.5
ipython==8.29.0
isoduration==20.11.0
jaraco.collections==5.1.0
jaraco.context==5.3.0
jaraco.functools==4.0.1
jaraco.text==3.12.1
jedi==0.19.1
jieba==0.42.1
Jinja2
jiter==0.6.1
jmespath==1.0.1
joblib==1.4.2
json5==0.9.25
jsonlines==4.0.0
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter-events==0.10.0
jupyter-lsp==2.2.5
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyter_server==2.14.2
jupyter_server_terminals==0.5.3
jupyterlab==4.3.0
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
kiwisolver==1.4.7
latex2mathml==3.77.0
lazy_loader==0.4
Levenshtein==0.26.0
librosa==0.10.2.post1
llvmlite==0.43.0
loguru
lxml==5.3.0
Markdown==3.7
markdown-it-py==3.0.0
markdown2==2.5.1
MarkupSafe
marshmallow==3.22.0
matplotlib==3.9.2
matplotlib-inline==0.1.7
mbstrdecoder==1.1.3
mdurl==0.1.2
mistune==3.0.2
mkl-service==2.4.0
mkl_fft 
mkl_random 
more-itertools==10.3.0
mpmath 
msgpack==1.1.0
multidict==6.1.0
multiprocess==0.70.16
mypy-extensions==1.0.0
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
nest-asyncio==1.6.0
networkx
ninja==1.11.1.1
nltk==3.9.1
nodeenv==1.9.1
notebook_shim==0.2.4
numba==0.60.0
numexpr==2.10.1
numpy==1.26.0
nvitop==1.5.0
objaverse==0.1.7
openai==1.51.2
opencv-python==4.10.0.84
openpyxl==3.1.5
orjson==3.10.7
overrides==7.7.0
packaging==24.1
pandas==2.2.3
pandocfilters==1.5.1
parameterized==0.9.0
parso==0.8.4
pathspec==0.12.1
pathvalidate==3.2.1
peft==0.12.0
pexpect==4.9.0
pillow
platformdirs==4.3.6
plyfile==1.1
polygon==1.2.6
pooch==1.8.2
portalocker==2.10.1
pre_commit==4.0.1
prometheus_client==0.21.0
prompt_toolkit==3.0.48
propcache==0.2.0
protobuf==5.28.2
psutil==6.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
py-cpuinfo==9.0.0
pyarrow==17.0.0
pybind11==2.13.6
pycocoevalcap==1.2
pycocotools==2.0.8
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
pydub==0.25.1
Pygments==2.18.0
pyparsing==3.2.0
PySocks
pytablewriter==1.2.0
python-dateutil==2.9.0.post0
python-json-logger==2.0.7
python-multipart==0.0.12
pytorchvideo==0.1.5
pytz==2024.2
PyYAML==6.0.2
pyzmq==26.2.0
RapidFuzz==3.10.0
referencing==0.35.1
regex==2024.7.24
requests
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.9.2
rouge-chinese==1.0.3
rouge_score==0.1.2
rpds-py==0.20.1
ruff==0.6.9
s3transfer==0.10.2
sacrebleu==2.4.3
safetensors==0.5.3
scikit-learn==1.5.2
scipy==1.14.1
semantic-version==2.10.0
Send2Trash==1.8.3
sentencepiece==0.2.0
shellingham==1.5.4
shortuuid==1.0.13
shtab==1.7.1
six==1.16.0
sniffio==1.3.1
soundfile==0.12.1
soupsieve==2.6
soxr==0.5.0.post1
sqlitedict==2.1.0
sse-starlette==2.1.3
stack-data==0.6.3
starlette==0.39.2
svgwrite==1.4.3
sympy 
tabledata==1.3.3
tabulate==0.9.0
tcolorpy==0.1.6
tenacity==9.0.0
tensorboard==2.18.0
tensorboard-data-server==0.7.2
termcolor==2.5.0
terminado==0.18.1
threadpoolctl==3.5.0
tiktoken==0.8.0
timm==1.0.10
tinycss2==1.4.0
tokenizers==0.21.1
tomli==2.0.2
tomlkit==0.12.0
torch==2.1.2
torchaudio==2.1.2
torchsummary==1.5.1
torchvision==0.15.2
tornado==6.4.1
tqdm==4.66.5
tqdm-multiprocess==0.0.11
traitlets==5.14.3
transformers==4.51.3
transformers-stream-generator==0.0.5
triton==2.1.0
trl==0.9.6
typeguard==4.3.0
typepy==1.3.2
typer==0.12.5
types-python-dateutil==2.9.0.20241003
typing-inspect==0.9.0
typing_extensions==4.12.2
tyro==0.8.14
tzdata==2024.2
uri-template==1.3.0
urllib3
uvicorn==0.31.1
virtualenv==20.26.6
wavedrom==2.0.3.post3
wcwidth==0.2.13
webcolors==24.8.0
webencodings==0.5.1
websocket-client==1.8.0
websockets==12.0
Werkzeug==3.0.4
xxhash==3.5.0
yacs==0.1.8
yarl==1.15.1
zhconv==1.4.3
zipp==3.19.2
zss==1.2.0
zstandard==0.23.0
